[modes] use 1:1 ServerInstances
