+++ mitmproxy/proxy/layers/modes.py
@@ -2,16 +2,14 @@ import socket
 import struct
 from abc import ABCMeta
 from dataclasses import dataclass
-from typing import TYPE_CHECKING, cast, Optional
+from typing import Optional
 
 from mitmproxy import connection
 from mitmproxy.proxy import commands, events, layer
 from mitmproxy.proxy.commands import StartHook
 from mitmproxy.proxy.layers import dns, tls
+from mitmproxy.proxy.mode_specs import ReverseMode
 from mitmproxy.proxy.utils import expect
-
-if TYPE_CHECKING:
-    from mitmproxy.proxy import mode_specs
 
 
 class HttpProxy(layer.Layer):
@@ -57,7 +55,8 @@ class DestinationKnown(layer.Layer, metaclass=ABCMeta):
 class ReverseProxy(DestinationKnown):
     @expect(events.Start)
     def _handle_event(self, event: events.Event) -> layer.CommandGenerator[None]:
-        spec = cast("mode_specs.ReverseMode", self.context.client.proxy_mode)
+        spec = self.context.client.proxy_mode
+        assert isinstance(spec, ReverseMode)
         self.context.server.address = spec.address
 
         if spec.scheme == "https" or spec.scheme == "tls" or spec.scheme == "dtls":
