+++ test/mitmproxy/addons/test_upstream_auth.py
@@ -2,7 +2,7 @@ import base64
 import pytest
 
 from mitmproxy import exceptions
-from mitmproxy.proxy.mode import ProxyMode
+from mitmproxy.proxy.mode_specs import ProxyMode
 from mitmproxy.test import taddons
 from mitmproxy.test import tflow
 from mitmproxy.addons import upstream_auth
